# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: A3 Partner Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}:
    get:
      operationId: getEvent
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Event, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
    patch:
      operationId: updateEvent
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
    delete:
      operationId: deleteEvent
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Deleted, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
...
  /billing/orders/{id}/resolve:
    get:
      operationId: resolveBilling
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Resolved, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
  /billing/orders/{id}/override:
    post:
      operationId: overrideBilling
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Updated order, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
...
components:
  schemas:
...
    AdminGenericObject:
      type: object
      additionalProperties: true
...